Restructure atl and cgs.
