+++ atl-checker/src/algorithms/certain_zero/search_strategy/mod.rs
@@ -1,4 +1,4 @@
-use crate::atl::atl_cgs_edg::{Edge, NegationEdge, Vertex};
+use crate::atl_cgs_edg::{Edge, NegationEdge, Vertex};
 
 pub mod bfs;
 pub mod dfs;
