+++ atl-checker/src/gamestructure/lcgs/ir/mod.rs
