+++ atl-checker/src/gamestructure/lcgs/mod.rs
